refactor 4: update thuộc tính version Long -> int

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/ApplicationApprovalImpl_New.java
@@ -85,7 +85,7 @@ public class ApplicationApprovalImpl_New implements ApplicationApprovalService_N
 	}
 
 	@Override
-	public void delete(String companyID, String appID, Long version, ApplicationType_Old appType) {
+	public void delete(String companyID, String appID, int version, ApplicationType_Old appType) {
 		switch (appType) {
 		case STAMP_APPLICATION:
 			appStampRepository.delete(companyID, appID);
