+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/after/AfterProcessDeleteImpl.java
@@ -50,7 +50,7 @@ public class AfterProcessDeleteImpl implements AfterProcessDelete {
 	private InterimRemainDataMngRegisterDateChange interimRemainDataMngRegisterDateChange;
 	
 	@Override
-	public ProcessDeleteResult screenAfterDelete(String companyID,String appID, Long version) {
+	public ProcessDeleteResult screenAfterDelete(String companyID,String appID, int version) {
 		boolean isProcessDone = true;
 		boolean isAutoSendMail = false;
 		List<String> autoSuccessMail = new ArrayList<>();
