+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/before/DetailBeforeUpdate.java
@@ -23,12 +23,12 @@ public interface DetailBeforeUpdate {
 	 * @param wkTimeCd
 	 */
 	public void processBeforeDetailScreenRegistration(String companyID, String employeeID, GeneralDate appDate, int employeeRouteAtr, String appID, 
-			PrePostAtr postAtr, Long version,String wkTypeCD, String wkTimeCd);
+			PrePostAtr postAtr, int version,String wkTypeCD, String wkTimeCd);
 	
 	/**
 	 * 1.排他チェック
 	 */
-	public void exclusiveCheck(String companyID, String appID, Long version);
+	public void exclusiveCheck(String companyID, String appID, int version);
 	/**
 	 * 4-1.詳細画面登録前の処理 (CMM045)
 	 * @param companyID 会社ID 
@@ -39,7 +39,7 @@ public interface DetailBeforeUpdate {
 	 * @param postAtr 事前事後区分
 	 */
 	public boolean processBefDetailScreenReg(String companyID, String employeeID, GeneralDate appDate, int employeeRouteAtr, String appID, 
-			PrePostAtr_Old postAtr, Long version);
+			PrePostAtr_Old postAtr, int version);
 	/**
 	 * 1.排他チェック (CMM045)
 	 * @param companyID
@@ -47,7 +47,7 @@ public interface DetailBeforeUpdate {
 	 * @param version
 	 * @return
 	 */
-	public boolean exclusiveCheckErr(String companyID, String appID, Long version);
+	public boolean exclusiveCheckErr(String companyID, String appID, int version);
 
 	/**
 	 * 勤務種類、就業時間帯チェックのメッセージを表示
